refactor: move run text to summary displayer
