+++ internal/pkg/client/run_summary.go
@@ -11,6 +11,7 @@ import (
 	"net/http"
 	"sort"
 	"strings"
+	"time"
 
 	"github.com/hashicorp/go-tfe/v2/api/models"
 	"github.com/microsoft/kiota-abstractions-go/serialization"
@@ -29,6 +30,11 @@ type RunSummary struct {
 	PolicyCheckStatus string             `json:"policy_check_status,omitempty"` // "hard_failed", "soft_failed", "errored"
 	PolicyEvaluations []PolicyEvalResult `json:"policy_evaluations,omitempty"`
 	TaskResults       []TaskResult       `json:"task_results,omitempty"`
+	// RunURL is the HCP Terraform UI URL for the run. Set by the caller when known.
+	RunURL string `json:"run_url,omitempty"`
+	// Elapsed is the time from run creation to terminal status, derived from
+	// the run's status timestamps. Zero if timestamps are unavailable.
+	Elapsed uint64 `json:"elapsed_seconds,omitempty"`
 }
 
 // PolicyEvalResult holds the outcome of a policy evaluation (OPA/Sentinel via task stages).
@@ -109,12 +115,58 @@ func NewRunSummary(ctx context.Context, c *Client, runID string) (*RunSummary, e
 		return nil, fmt.Errorf("fetching run %s: %w", runID, err)
 	}
 
-	status := run.GetData().GetAttributes().GetStatus()
+	attrs := run.GetData().GetAttributes()
+	status := attrs.GetStatus()
 	if status == nil {
 		return nil, fmt.Errorf("run %s has no status", runID)
 	}
 
-	return buildRunSummary(ctx, c, runID, *status)
+	result, err := buildRunSummary(ctx, c, runID, *status)
+	if err != nil {
+		return nil, err
+	}
+
+	result.Elapsed = uint64(elapsedFromTimestamps(attrs.GetCreatedAt(), attrs.GetStatusTimestamps(), *status).Seconds())
+	return result, nil
+}
+
+// elapsedFromTimestamps derives the run duration from the terminal status timestamp
+// minus the run's created-at time. Returns zero if either value is unavailable.
+func elapsedFromTimestamps(createdAt *time.Time, ts models.Runs_attributes_statusTimestampsable, status models.Runs_attributes_status) time.Duration {
+	if createdAt == nil || ts == nil {
+		return 0
+	}
+
+	var endTime *time.Time
+	switch status {
+	case models.APPLIED_RUNS_ATTRIBUTES_STATUS:
+		endTime = ts.GetAppliedAt()
+	case models.PLANNED_AND_FINISHED_RUNS_ATTRIBUTES_STATUS:
+		endTime = ts.GetPlannedAndFinishedAt()
+	case models.PLANNED_AND_SAVED_RUNS_ATTRIBUTES_STATUS:
+		endTime = ts.GetPlannedAndSavedAt()
+	case models.PLANNED_RUNS_ATTRIBUTES_STATUS:
+		endTime = ts.GetPlannedAt()
+	case models.ERRORED_RUNS_ATTRIBUTES_STATUS:
+		endTime = ts.GetErroredAt()
+	case models.CANCELED_RUNS_ATTRIBUTES_STATUS:
+		endTime = ts.GetCanceledAt()
+	case models.DISCARDED_RUNS_ATTRIBUTES_STATUS:
+		endTime = ts.GetDiscardedAt()
+	case models.POLICY_SOFT_FAILED_RUNS_ATTRIBUTES_STATUS:
+		endTime = ts.GetPolicySoftFailedAt()
+	case models.POLICY_OVERRIDE_RUNS_ATTRIBUTES_STATUS:
+		endTime = ts.GetPolicyCheckedAt()
+	}
+
+	if endTime == nil || endTime.IsZero() {
+		return 0
+	}
+	d := endTime.Sub(*createdAt)
+	if d < 0 {
+		return 0
+	}
+	return d.Round(time.Second)
 }
 
 func buildRunSummary(ctx context.Context, c *Client, runID string, status models.Runs_attributes_status) (*RunSummary, error) {
